+++ ios-technical-test-destinationguide/DestinationGuide/View/DetailsController.swift
@@ -1,15 +1,17 @@
 //
-//  DestinationDetailsController.swift
-//  DestinationDetailsController
+//  DetailsController.swift
+//  DetailsController
 //
-//  Created by Kel_Jellysh on 15/08/2021.
+//  Created by Kel_Jellysh on 17/08/2021.
 //
 
 import UIKit
 import WebKit
-class DestinationDetailsController: UIViewController {
 
+class DetailsController: UINavigationController, WKNavigationDelegate {
+    
     let identifier = "WebPage"
+    
     var webPage: WKWebView = {
         let web = WKWebView()
         web.backgroundColor = .white
@@ -19,10 +21,16 @@ class DestinationDetailsController: UIViewController {
     
     var idDestination: String?
     var travelDetails: [DestinationDetails]?
+    
+    override func loadView() {
+        webPage.navigationDelegate = self
+        view = webPage
+    }
    
     override func viewDidLoad() {
         super.viewDidLoad()
         setWeb()
+       
       
         // Do any additional setup after loading the view.
         DestinationFetchingService.shared.getDestinationDetails(for: idDestination ?? "") { result in
@@ -35,18 +43,20 @@ class DestinationDetailsController: UIViewController {
                 return
             }
         }
-        
+        //déclarer l'URL ici
+//        let url = URL(string: travelDetails[index].url)!
+//        webPage.load(URLRequest(url: url))
+        webPage.allowsBackForwardNavigationGestures = true
     }
+
     
-    private func setWeb() {
+    func setWeb() {
         view.addSubview(webPage)
         webPage.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: 0).isActive = true
         webPage.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 0).isActive = true
         webPage.topAnchor.constraint(equalTo: view.topAnchor, constant: 0).isActive = true
         webPage.bottomAnchor.constraint(equalTo: view.bottomAnchor, constant: 0).isActive = true
     }
-    
-    
 
     /*
     // MARK: - Navigation
