+++ ios-technical-test-destinationguide/DestinationGuide/View/Extension/ExtensionUIImageView.swift
@@ -1,0 +1,34 @@
+//
+//  ExtensionUIImage.swift
+//  ExtensionUIImage
+//
+//  Created by Kel_Jellysh on 15/08/2021.
+//
+
+import Foundation
+import UIKit
+
+extension UIImageView {
+    
+    
+    func downloaded(from url: URL, contentMode mode: UIView.ContentMode = .scaleAspectFill) {
+        contentMode = mode
+        URLSession.shared.dataTask(with: url) { [weak self] data, response, error in
+            guard
+                let httpURLResponse = response as? HTTPURLResponse, httpURLResponse.statusCode == 200,
+                let mimeType = response?.mimeType, mimeType.hasPrefix(""),
+                let data = data, error == nil,
+                let image = UIImage(data: data)
+            else { return }
+            DispatchQueue.main.async() {
+                self?.image = image
+            }
+        }.resume()
+    }
+    
+    func downloaded(from link: String, contentMode mode: UIView.ContentMode = .scaleAspectFit) {
+        guard let url = URL(string: link) else { return }
+        downloaded(from: url, contentMode: mode)
+    }
+    
+}
